Adicionar correcoes simples de array
