+++ MeuDevocionalWidget/WidgetContent.swift
@@ -1,0 +1,37 @@
+//
+//  WidgetContent.swift
+//  MeuDevocional
+//
+//  Created by Beatriz Duque on 19/01/22.
+//
+
+import Foundation
+import UIKit
+
+///array com conteudos dos widgets
+var widgetsTitle: [String] = ["Meu Devocional",
+                            "Gratidão",
+                            "Worship Time!",
+                            "Salmo 37:5",
+                            "Mateus 6:34",
+                            "João 3:16",
+                            "Momento de oração",
+                            "Mateus 11:21",
+                            "Filipenses 1:9",
+                            "João 8:32",
+                            "2 Coríntios 5:7",
+                            "Lucas 12:34",
+                            "1 João 4:8"]
+var widgetsContent: [String] = ["Não esqueça de anotar e compartilhar a devocional de hoje!",
+                              "Já adicionou um motivo para ser grato no Mural hoje?",
+                              "Que tal escolher uma devocional para leitura e louvor hoje?",
+                              "'Entrega o teu caminho ao Senhor; confia nele, e ele tudo fará.'",
+                              "Portanto, não vos inquieteis com o dia de amanhã [...]",
+                              "'Acima de tudo, porém, revistam-se do amor, que é o elo perfeito.'",
+                              "Conta pra Ele, conte com Ele!",
+                              "'Vinde a mim todos os que estais cansados, e eu vos aliviarei.'",
+                              "E peço isto: que o vosso amor cresça mais e mais [...]",
+                              "E conhecereis a verdade, e a verdade vos libertará.",
+                              "Porque vivemos por fé, e não pelo que vemos.",
+                              "Porque, onde estiver o vosso tesouro, ali estará o vosso coração.",
+                              "Quem não ama não conhece a Deus, porque Deus é amor."]
